requirements.txt
keyboard==0.13.5
packaging==24.1
pillow==11.0.0
pytesseract==0.3.13
pyperclip